chore: modularize utilities
