+++ jaguar-express/src/utils/device.ts
@@ -1,0 +1,30 @@
+// Verificar si es dispositivo móvil
+export function isMobile(): boolean {
+  return /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
+    navigator.userAgent
+  );
+}
+
+// Verificar si es iOS
+export function isIOS(): boolean {
+  return /iPad|iPhone|iPod/.test(navigator.userAgent);
+}
+
+// Verificar si es Android
+export function isAndroid(): boolean {
+  return /Android/.test(navigator.userAgent);
+}
+
+// Obtener información del dispositivo
+export function getDeviceInfo() {
+  return {
+    isMobile: isMobile(),
+    isIOS: isIOS(),
+    isAndroid: isAndroid(),
+    userAgent: navigator.userAgent,
+    platform: navigator.platform,
+    language: navigator.language,
+    cookieEnabled: navigator.cookieEnabled,
+    onLine: navigator.onLine,
+  };
+}
