+++ jaguar-express/src/utils/format.ts
@@ -1,0 +1,146 @@
+import { type ClassValue, clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
+
+// Utility para combinar clases de Tailwind
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
+
+// Formateo de moneda
+export function formatCurrency(amount: number): string {
+  return new Intl.NumberFormat('es-CO', {
+    style: 'currency',
+    currency: 'COP',
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 0,
+  }).format(amount);
+}
+
+// Formateo de números
+export function formatNumber(num: number): string {
+  return new Intl.NumberFormat('es-CO').format(num);
+}
+
+// Formateo de fechas
+export function formatDate(
+  date: string | Date,
+  options?: Intl.DateTimeFormatOptions
+): string {
+  const dateObj = typeof date === 'string' ? new Date(date) : date;
+
+  const defaultOptions: Intl.DateTimeFormatOptions = {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  };
+
+  return new Intl.DateTimeFormat('es-CO', { ...defaultOptions, ...options }).format(
+    dateObj
+  );
+}
+
+// Formateo de tiempo relativo
+export function formatRelativeTime(date: string | Date): string {
+  const dateObj = typeof date === 'string' ? new Date(date) : date;
+  const now = new Date();
+  const diffInSeconds = Math.floor((now.getTime() - dateObj.getTime()) / 1000);
+
+  if (diffInSeconds < 60) {
+    return 'Hace un momento';
+  }
+
+  const diffInMinutes = Math.floor(diffInSeconds / 60);
+  if (diffInMinutes < 60) {
+    return `Hace ${diffInMinutes} minuto${diffInMinutes > 1 ? 's' : ''}`;
+  }
+
+  const diffInHours = Math.floor(diffInMinutes / 60);
+  if (diffInHours < 24) {
+    return `Hace ${diffInHours} hora${diffInHours > 1 ? 's' : ''}`;
+  }
+
+  const diffInDays = Math.floor(diffInHours / 24);
+  if (diffInDays < 7) {
+    return `Hace ${diffInDays} día${diffInDays > 1 ? 's' : ''}`;
+  }
+
+  return formatDate(dateObj);
+}
+
+// Formateo de tiempo de entrega
+export function formatDeliveryTime(minutes: number): string {
+  if (minutes < 60) {
+    return `${minutes} min`;
+  }
+
+  const hours = Math.floor(minutes / 60);
+  const remainingMinutes = minutes % 60;
+
+  if (remainingMinutes === 0) {
+    return `${hours} hora${hours > 1 ? 's' : ''}`;
+  }
+
+  return `${hours}h ${remainingMinutes}min`;
+}
+
+// Formateo de teléfono
+export function formatPhone(phone: string): string {
+  const cleaned = phone.replace(/\D/g, '');
+
+  if (cleaned.startsWith('57')) {
+    const number = cleaned.substring(2);
+    return `+57 ${number.substring(0, 3)} ${number.substring(3, 6)} ${number.substring(6)}`;
+  }
+
+  if (cleaned.length === 10) {
+    return `${cleaned.substring(0, 3)} ${cleaned.substring(3, 6)} ${cleaned.substring(6)}`;
+  }
+
+  return phone;
+}
+
+// Sanitización de texto
+export function sanitizeText(text: string): string {
+  return text.trim().replace(/\s+/g, ' ');
+}
+
+// Truncar texto
+export function truncateText(text: string, maxLength: number): string {
+  if (text.length <= maxLength) return text;
+  return text.substring(0, maxLength).trim() + '...';
+}
+
+// Capitalizar primera letra
+export function capitalize(text: string): string {
+  return text.charAt(0).toUpperCase() + text.slice(1).toLowerCase();
+}
+
+// Capitalizar cada palabra
+export function capitalizeWords(text: string): string {
+  return text.split(' ').map(word => capitalize(word)).join(' ');
+}
+
+// Generar slug
+export function generateSlug(text: string): string {
+  return text
+    .toLowerCase()
+    .normalize('NFD')
+    .replace(/[\u0300-\u036f]/g, '') // Remover acentos
+    .replace(/[^a-z0-9\s-]/g, '') // Remover caracteres especiales
+    .replace(/\s+/g, '-') // Reemplazar espacios con guiones
+    .replace(/-+/g, '-') // Remover guiones múltiples
+    .trim();
+}
+
+// Formatear bytes
+export function formatBytes(bytes: number, decimals: number = 2): string {
+  if (bytes === 0) return '0 Bytes';
+
+  const k = 1024;
+  const dm = decimals < 0 ? 0 : decimals;
+  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
+}
